split pagepiling init
